Command Cursors, Player & UI Cleanup
Made commands change the cursor to reflect targeting, separated UI code from Player object and incorporated input to the UI Controller. Started on getting the cursor to update as you hover the mouse over objects.
