+++ Assets/Player/CursorSprite.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 404e224376d95f349ab30c7910ad568f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
