+++ Assets/Player/Player.cs
@@ -1,6 +1,7 @@
 ﻿using MarsTS.Events;
 using MarsTS.Players.Input;
 using MarsTS.Players.Teams;
+using MarsTS.UI;
 using MarsTS.Units;
 using MarsTS.Units.Commands;
 using MarsTS.World;
@@ -39,85 +40,25 @@ namespace MarsTS.Players {
 		public static Agent EventAgent { get { return instance.eventAgent; } }
 		private Agent eventAgent;
 
+		public static UIController UI { get { return instance.uiController; } }
+		private UIController uiController;
+
 		[SerializeField]
 		private float cameraSpeed;
 
 		private Vector2 moveDirection;
-		
-
-		[SerializeField]
-		private RectTransform selectionSquare;
-		[SerializeField]
-		private GameObject selectionPrefab;
-		private bool mouseHeld = false;
-		private Vector2 drawStart;
-		private Vector2 drawMouse;
 
 		private void Awake () {
 			instance = this;
 			view = GetComponent<Camera>();
 			inputController = GetComponent<InputHandler>();
 			eventAgent = GetComponent<Agent>();
+			uiController = GetComponent<UIController>();
 			alternate = false;
-		}
-
-		private void Start () {
-			
 		}
 
 		private void Update () {
 			transform.position = transform.position + (cameraSpeed * Time.deltaTime * new Vector3(moveDirection.x, 0, moveDirection.y));
-
-			if (mouseHeld) {
-				drawMouse = cursorPos;
-
-				if (!selectionSquare.gameObject.activeInHierarchy && Vector2.Distance(drawStart, drawMouse) > 2f) selectionSquare.gameObject.SetActive(true);
-
-				Vector2 topLeft;
-				Vector2 bottomRight;
-
-				topLeft.x = drawMouse.x >= drawStart.x ? drawStart.x : drawMouse.x;
-				topLeft.y = drawMouse.y >= drawStart.y ? drawMouse.y : drawStart.y;
-				bottomRight.x = drawMouse.x >= drawStart.x ? drawMouse.x : drawStart.x;
-				bottomRight.y = drawMouse.y >= drawStart.y ? drawStart.y : drawMouse.y;
-
-				selectionSquare.offsetMin = new Vector2(topLeft.x, bottomRight.y);
-				selectionSquare.offsetMax = new Vector2(bottomRight.x - Screen.width, topLeft.y - Screen.height);
-			}
-		}
-
-		private void EndSelectionDraw () {
-			if (Vector2.Distance(drawStart, drawMouse) > 2f) {
-				Vector3[] screenCorners = new Vector3[4];
-
-				selectionSquare.GetWorldCorners(screenCorners);
-
-				Vector3[] orderedCorners = new Vector3[4];
-
-				Vector3[] rayVerts = new Vector3[4];
-				Vector3[] dirVerts = new Vector3[4];
-
-				orderedCorners[0] = screenCorners[1];
-				orderedCorners[1] = screenCorners[2];
-				orderedCorners[2] = screenCorners[0];
-				orderedCorners[3] = screenCorners[3];
-
-				for (int i = 0; i < orderedCorners.Length; i++) {
-					Ray ray = view.ScreenPointToRay(orderedCorners[i]);
-
-					if (Physics.Raycast(ray, out RaycastHit hit, 50000f)) {
-						rayVerts[i] = hit.point;
-						dirVerts[i] = ray.origin - hit.point;
-						//Debug.DrawLine(view.ScreenToWorldPoint(orderedCorners[i]), hit.point, Color.red, 1.0f);
-					}
-				}
-
-				MeshCollider collider = Instantiate(selectionPrefab).GetComponent<MeshCollider>();
-				collider.sharedMesh = SelectionMesh(rayVerts, dirVerts);
-				Destroy(collider.gameObject, 0.2f);
-			}
-
-			selectionSquare.gameObject.SetActive(false);
 		}
 
 		/*	Input Functions	*/
@@ -131,39 +72,15 @@ namespace MarsTS.Players {
 		}
 
 		public void Select (InputAction.CallbackContext context) {
-			switch (context.phase) {
-				//Press Down
-				case InputActionPhase.Started: {
-					if (Input.HoveringUI) return;
-					drawStart = MousePos;
-					mouseHeld = true;
-					break;
-				}
+			if (context.phase == InputActionPhase.Canceled) {
+				if (UI.Hovering) return;
+				if (!Include) ClearSelection();
 
-				//Release
-				//If we've reached the held threshold, then we want to select everything within the drawn box
-				//Otherwise its a normal click
-				case InputActionPhase.Canceled: {
-					if (selectionSquare.gameObject.activeInHierarchy) {
-						if (!Include) ClearSelection();
-						EndSelectionDraw();
-						mouseHeld = false;
-						return;
-					}
-					
-					if (Input.HoveringUI) return;
-					if (!Include) ClearSelection();
+				Ray ray = ViewPort.ScreenPointToRay(cursorPos);
 
-					Ray ray = ViewPort.ScreenPointToRay(cursorPos);
-
-					if (Physics.Raycast(ray, out RaycastHit hit, 1000f, GameWorld.SelectableMask)) {
-						Unit hitUnit = hit.collider.gameObject.GetComponentInParent<ISelectable>().Get();
-						SelectUnit(hitUnit);
-					}
-
-					EventBus.Global(new SelectEvent(true, Selected));
-					mouseHeld = false;
-					break;
+				if (Physics.Raycast(ray, out RaycastHit hit, 1000f, GameWorld.SelectableMask)) {
+					Unit hitUnit = hit.collider.gameObject.GetComponentInParent<ISelectable>().Get();
+					SelectUnit(hitUnit);
 				}
 			}
 		}
@@ -182,9 +99,11 @@ namespace MarsTS.Players {
 					target.Select(true);
 				}
 			}
+
+			EventBus.Global(new SelectEvent(Selected));
 		}
 
-		private void ClearSelection () {
+		public void ClearSelection () {
 			foreach (Roster units in selected.Values) {
 				foreach (Unit unit in units.List()) {
 					unit.Select(false);
@@ -194,6 +113,7 @@ namespace MarsTS.Players {
 			}
 
 			selected.Clear();
+			EventBus.Global(new SelectEvent(Selected));
 		}
 
 		private Roster GetRoster (string type) {
@@ -203,7 +123,7 @@ namespace MarsTS.Players {
 		}
 
 		public void Command (InputAction.CallbackContext context) {
-			if (context.canceled && !Input.HoveringUI) {
+			if (context.canceled && !UI.Hovering) {
 				Ray ray = ViewPort.ScreenPointToRay(cursorPos);
 
 				Physics.Raycast(ray, out RaycastHit walkableHit, 1000f, GameWorld.WalkableMask);
@@ -236,25 +156,6 @@ namespace MarsTS.Players {
 		public void Alternate (InputAction.CallbackContext context) {
 			if (context.performed) alternate = true;
 			if (context.canceled) alternate = false;
-		}
-
-		private Mesh SelectionMesh (Vector3[] corners, Vector3[] vecs) {
-			Vector3[] verts = new Vector3[8];
-			int[] tris = { 0, 1, 2, 2, 1, 3, 4, 6, 0, 0, 6, 2, 6, 7, 2, 2, 7, 3, 7, 5, 3, 3, 5, 1, 5, 0, 1, 1, 4, 0, 4, 5, 6, 6, 5, 7 }; //map the tris of our cube
-
-			for (int i = 0; i < 4; i++) {
-				verts[i] = corners[i];
-			}
-
-			for (int j = 4; j < 8; j++) {
-				verts[j] = corners[j - 4] + vecs[j - 4];
-			}
-
-			Mesh selectionMesh = new Mesh();
-			selectionMesh.vertices = verts;
-			selectionMesh.triangles = tris;
-
-			return selectionMesh;
 		}
 
 		private void OnTriggerEnter (Collider other) {
