+++ Assets/Player/UIController.cs
@@ -2,11 +2,15 @@ using MarsTS.Events;
 using MarsTS.Players;
 using MarsTS.Units;
 using MarsTS.Units.Commands;
+using MarsTS.World;
 using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.EventSystems;
 using UnityEngine.InputSystem;
+using UnityEngine.UI;
+using static UnityEditor.PlayerSettings.SplashScreen;
 
 namespace MarsTS.UI {
 
@@ -19,6 +23,21 @@ namespace MarsTS.UI {
 		private string[] profileIndex;
 
 		private int primaryIndex;
+
+		[SerializeField]
+		private RectTransform selectionSquare;
+		[SerializeField]
+		private GameObject selectionPrefab;
+
+		private bool mouseHeld = false;
+		private Vector2 drawStart;
+		private Vector2 drawMouse;
+
+		[SerializeField]
+		private CursorSprite defaultCursor;
+
+		[SerializeField]
+		private CursorSprite defaultCursorSprite;
 
 		private string PrimarySelected {
 			get {
@@ -39,14 +58,144 @@ namespace MarsTS.UI {
 
 		private string primarySelected;
 
+		[SerializeField]
+		private GraphicRaycaster canvasRaycaster;
+
+		public bool Hovering {
+			get {
+				Vector2 mousePos = Player.MousePos;
+
+				PointerEventData pointData = new PointerEventData(null);
+				pointData.position = mousePos;
+				List<RaycastResult> results = new List<RaycastResult>();
+
+				canvasRaycaster.Raycast(pointData, results);
+
+				return results.Count > 0;
+			}
+		}
+
 		private void Awake () {
 			commandProfiles = new();
+			ResetCursor();
 		}
 
 		private void Start () {
 			commandPanel = GameObject.Find("Command Zone").GetComponent<CommandPanel>();
 			unitPane = GameObject.Find("Unit Pane").GetComponent<UnitPane>();
 			EventBus.AddListener<SelectEvent>(OnSelection);
+		}
+
+		private void Update () {
+			if (mouseHeld) {
+				drawMouse = Player.MousePos;
+
+				if (!selectionSquare.gameObject.activeInHierarchy && Vector2.Distance(drawStart, drawMouse) > 2f) selectionSquare.gameObject.SetActive(true);
+				Vector2 topLeft;
+				Vector2 bottomRight;
+
+				topLeft.x = drawMouse.x >= drawStart.x ? drawStart.x : drawMouse.x;
+				topLeft.y = drawMouse.y >= drawStart.y ? drawMouse.y : drawStart.y;
+				bottomRight.x = drawMouse.x >= drawStart.x ? drawMouse.x : drawStart.x;
+				bottomRight.y = drawMouse.y >= drawStart.y ? drawStart.y : drawMouse.y;
+
+				selectionSquare.offsetMin = new Vector2(topLeft.x, bottomRight.y);
+				selectionSquare.offsetMax = new Vector2(bottomRight.x - Screen.width, topLeft.y - Screen.height);
+			}
+		}
+
+		public void ResetCursor () {
+			Cursor.SetCursor(defaultCursor.texture, defaultCursor.target, CursorMode.Auto);
+		}
+
+		public void Select (InputAction.CallbackContext context) {
+			switch (context.phase) {
+				//Press Down
+				case InputActionPhase.Started: {
+					if (Hovering) return;
+					drawStart = Player.MousePos;
+					mouseHeld = true;
+					break;
+				}
+
+				case InputActionPhase.Canceled: {
+					if (selectionSquare.gameObject.activeInHierarchy) {
+						if (!Player.Include) Player.Main.ClearSelection();
+						EndSelectionDraw();
+						mouseHeld = false;
+					}
+
+					break;
+				}
+			}
+		}
+
+		public void Look (InputAction.CallbackContext context) {
+			if (Hovering) return;
+
+			Vector2 mousePos = context.ReadValue<Vector2>();
+
+			Ray ray = Player.ViewPort.ScreenPointToRay(mousePos);
+
+			if (Physics.Raycast(ray, out RaycastHit selectable, 1000f, GameWorld.SelectableMask)) {
+				//Cursor.SetCursor();
+			}
+			else if (Physics.Raycast(ray, out RaycastHit ground, 1000f, GameWorld.WalkableMask)) {
+
+			}
+		}
+
+		private void EndSelectionDraw () {
+			if (Vector2.Distance(drawStart, drawMouse) > 2f) {
+				Vector3[] screenCorners = new Vector3[4];
+
+				selectionSquare.GetWorldCorners(screenCorners);
+
+				Vector3[] orderedCorners = new Vector3[4];
+
+				Vector3[] rayVerts = new Vector3[4];
+				Vector3[] dirVerts = new Vector3[4];
+
+				orderedCorners[0] = screenCorners[1];
+				orderedCorners[1] = screenCorners[2];
+				orderedCorners[2] = screenCorners[0];
+				orderedCorners[3] = screenCorners[3];
+
+				for (int i = 0; i < orderedCorners.Length; i++) {
+					Ray ray = Player.ViewPort.ScreenPointToRay(orderedCorners[i]);
+
+					if (Physics.Raycast(ray, out RaycastHit hit, 50000f)) {
+						rayVerts[i] = hit.point;
+						dirVerts[i] = ray.origin - hit.point;
+						//Debug.DrawLine(view.ScreenToWorldPoint(orderedCorners[i]), hit.point, Color.red, 1.0f);
+					}
+				}
+
+				MeshCollider collider = Instantiate(selectionPrefab).GetComponent<MeshCollider>();
+				collider.sharedMesh = SelectionMesh(rayVerts, dirVerts);
+				Destroy(collider.gameObject, 0.2f);
+			}
+
+			selectionSquare.gameObject.SetActive(false);
+		}
+
+		private Mesh SelectionMesh (Vector3[] corners, Vector3[] vecs) {
+			Vector3[] verts = new Vector3[8];
+			int[] tris = { 0, 1, 2, 2, 1, 3, 4, 6, 0, 0, 6, 2, 6, 7, 2, 2, 7, 3, 7, 5, 3, 3, 5, 1, 5, 0, 1, 1, 4, 0, 4, 5, 6, 6, 5, 7 }; //map the tris of our cube
+
+			for (int i = 0; i < 4; i++) {
+				verts[i] = corners[i];
+			}
+
+			for (int j = 4; j < 8; j++) {
+				verts[j] = corners[j - 4] + vecs[j - 4];
+			}
+
+			Mesh selectionMesh = new Mesh();
+			selectionMesh.vertices = verts;
+			selectionMesh.triangles = tris;
+
+			return selectionMesh;
 		}
 
 		private void OnSelection (SelectEvent _event) {
